Add seed to random num generator in SVDEuclidean
I know commits to master are frowned upon. But this
is important for future runs. Numpy documentation
claims that this is called upon initialization of
RandomState, which I can only assume is initialized
with import numpy. But I would MUCH rather be safe
and include it again, then sorry and have drastically
awful blending results in a few hours.